# openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/auth/register:
    post:
      operationId: register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input (see `code` and `details`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict with existing data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Create an account and sign in
      tags:
        - auth
  /api/v1/auth/login:
    post:
      operationId: login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input (see `code` and `details`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid or expired credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Sign in with e-mail and password
      tags:
        - auth
  /api/v1/auth/refresh:
    post:
      operationId: refreshTokens
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input (see `code` and `details`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid or expired credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Exchange a refresh token for a new token pair
      tags:
        - auth
  /api/v1/auth/logout:
    post:
      operationId: logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "204":
          description: ""
        "400":
          description: Invalid input (see `code` and `details`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      summary: Revoke a refresh token
      tags:
        - auth
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Missing, invalid or expired credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - bearer: []
      summary: Profile of the signed-in user
      tags:
        - auth
    delete:
      operationId: deleteAccount
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
      responses:
        "204":
          description: ""
        "400":
          description: Invalid input (see `code` and `details`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid or expired credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - bearer: []
      summary: Delete the account and the whole collection
      tags:
        - auth
  /api/v1/auth/change-password:
    post:
      description: Signs out all sessions (every refresh token is revoked) and returns
        a new token pair for the current device. Fails with
        `INVALID_CURRENT_PASSWORD` when the current password is wrong.
      operationId: changePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input (see `code` and `details`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid or expired credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - bearer: []
      summary: Change password
      tags:
        - auth
...
info:
  title: Game Shelf API
  description: >-
    REST API of Game Shelf – a catalogue for collectors of computer and console
    games.


    Authenticate with `POST /api/v1/auth/login` or `/register` and send the
    access token as `Authorization: Bearer <token>`. When it expires (401),
    exchange the refresh token at `POST /api/v1/auth/refresh`.


    Every error has the shape of `ErrorResponse`; branch on its `code`.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: collector@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          example: super-secret-123
        displayName:
          type: string
          nullable: true
          maxLength: 100
          example: Retro Rob
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          example: collector@example.com
        displayName:
          type: string
          nullable: true
          example: Retro Rob
        createdAt:
          format: date-time
          type: string
      required:
        - id
        - email
        - displayName
        - createdAt
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: "Short-lived JWT for the `Authorization: Bearer` header."
        expiresIn:
          type: integer
          description: Expiry of the access token in seconds.
          example: 900
        refreshToken:
          type: string
          description: Opaque single-use token for POST /auth/refresh. Every refresh
            returns a new one; reusing an old token revokes all sessions of the
            user.
        user:
          $ref: "#/components/schemas/User"
      required:
        - accessToken
        - expiresIn
        - refreshToken
        - user
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - INVALID_CREDENTIALS
        - INVALID_REFRESH_TOKEN
        - INVALID_CURRENT_PASSWORD
        - EMAIL_ALREADY_REGISTERED
        - GAME_NOT_FOUND
        - NOT_FOUND
        - TOO_MANY_REQUESTS
        - BAD_REQUEST
        - FORBIDDEN
        - CONFLICT
        - INTERNAL_ERROR
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        code:
          allOf:
            - $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          example: Validation failed
        details:
          description: Individual validation messages, present for VALIDATION_FAILED.
          example:
            - email must be an email
          type: array
          items:
            type: string
      required:
        - statusCode
        - code
        - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: collector@example.com
        password:
          type: string
          format: password
          example: super-secret-123
      required:
        - email
        - password
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 128
      required:
        - currentPassword
        - newPassword
    DeleteAccountRequest:
      type: object
      properties:
        password:
          type: string
          format: password
          description: Current password as a confirmation.
      required:
        - password
...